requirements.txt
google-generativeai>=0.8.0
pydantic>=2.5.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
rich>=13.7.0
click>=8.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
